requirements.txt
fastapi[standard]==0.115.6
matplotlib
networkx
openai
pydantic
python-dotenv
tiktoken
